refactor: extract useResolveBatchItem hook
